# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "土爾曼API文件",
        "description": "測試Swagger文件",
        "contact": {
            "email": "r567tw@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://rocky-stream-36893.herokuapp.com",
            "description": "Heroku"
        },
        {
            "url": "http://toolman.test",
            "description": "本機"
        }
    ],
    "paths": {
        "/api/ping": {
            "get": {
                "tags": [
                    "Ping"
                ],
                "summary": "Health checker",
                "description": "Health checker",
                "operationId": "ping",
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "取得全部文章",
                "description": "取得全部文章",
                "operationId": "getAllPosts",
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "建立文章",
                "description": "建立文章",
                "operationId": "createPost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "HelloWorld",
                                    "content": "HelloWolrd content"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Ping",
            "description": "ping"
        },
        {
            "name": "Post",
            "description": "post"
        }
    ]
}